requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlmodel==0.0.14
jinja2==3.1.2
python-dotenv==0.21.1
python-multipart==0.0.6
APScheduler==3.10.4
pytz==2023.3